websocket, dsl, kotlin 2.3.10, docs

+++ src/main/kotlin/io/github/xerooup/xhttp/request/Request.kt
@@ -8,7 +8,6 @@ import java.net.http.HttpResponse
 import java.time.Duration
 import kotlin.collections.iterator
 
-// request builder
 class Request internal constructor(
     private val method: Method,
     private val url: String
@@ -31,7 +30,7 @@ class Request internal constructor(
     }
 
     fun timeout(duration: kotlin.time.Duration): Request = apply {
-        timeoutMs = duration.inWholeMilliseconds.toLong()
+        timeoutMs = duration.inWholeMilliseconds
     }
 
     fun body(text: String, contentType: String): Request = apply {
